Add files via upload
Database for comic inventory. The purpose will be to search my collection for specific issues and find information about it, including writer, volume number, and condition. I will also be able to search an author and find which titles they have written, or search for a specific character and find the first appearance of the character.